Fixed unnecessary file read in getBurstOriginXYN
Originally when getBurstOriginXYN is called it calls h5read twice every time it is called to read the x and the y location. Since getBurstOriginXYN will be called by getAllBurstOriginXYN n amount of times depending on how many burst there are, the x and the y location should only be read once and passed in.